Added balancing algorithm, fixed start of program and some minor stuff IT WORKS FINALLY :)

+++ src/main/java/at/ac/tgm/hit/dezsys/hamplwortha/Calculate.java
@@ -1,5 +1,5 @@
 package at.ac.tgm.hit.dezsys.hamplwortha;
 
 public interface Calculate {
-    double calc(int iterations);
+    double calc(long iterations);
 }
