+++ src/main/java/at/ac/tgm/hit/dezsys/hamplwortha/CalculatePi.java
@@ -4,12 +4,12 @@ package at.ac.tgm.hit.dezsys.hamplwortha;
 public class CalculatePi implements Calculate {
 
     /**
-     * @see Calculate#calc(int)
+     * @see Calculate#calc(long)
      */
     @Override
-    public double calc(int iterations) {
+    public double calc(long iterations) {
         double res = 0;
-        for (int i = 1; i < iterations; i += 4) {
+        for (long i = 1; i < iterations; i += 4) {
             res += 1.0 / i - 1.0 / (i + 2);
         }
         return 4 * res;
